pyproject.toml
[tool.poetry]
name = "django27"
version = "0.1.0"
description = ""
authors = ["Glassyflute <yuliya.bobir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
psycopg2 = "^2.9.5"
Pillow = "^9.3.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
